--- oits-flask/shop_backend/choices/order_status.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class OrderStatus(Enum):
-    NEW = 'new'
-    IN_PROGRESS = 'in_progress'
-    COMPLETED = 'completed'
